座標関係を全てfloatに統一

+++ src/ofApp.h
@@ -33,17 +33,17 @@ private:
 	vector<ofVec2f> particle;
 	vector<ofVec2f> ptcl0;
 	//粒子
-	int dtc(int x_d, int y_d);
+	float dtc(float x_d, float y_d);
 
 	ofVec2f toRelativeC(ofVec2f v);
-	void toRelativeC(int* vx,int* vy);
+	void toRelativeC(float* vx, float* vy);
 	ofVec2f toAbsoluteC(ofVec2f &v);
-	void toAbsoluteC(int* vx, int* vy);
+	void toAbsoluteC(float* vx, float* vy);
 
-	int toRelativeCx(int vx);
-	int toRelativeCy(int vy);
-	int toAbsoluteCx(int vx);
-	int toAbsoluteCy(int vy);
+	float toRelativeCx(float vx);
+	float toRelativeCy(float vy);
+	float toAbsoluteCx(float vx);
+	float toAbsoluteCy(float vy);
 	vector<ofVec4f> rigidBox;
 		
 };
